- FINALLY finished height effects passability! I wrote a description of what I discovered KaM uses, it's a bit odd. But it matches exactly for the maps I've tested.  - WIP woodcutter choosing best direction, SHOULDN'T BE INCLUDED IN DEMO AS IT IS NOW!!! Uncomment the two lines in Terrain that restore it to previous method if you are going release. Instructions are included with @Krom tag.  - A few more shortcuts added  - Updated Readme/todo list a bit  - Some testing, no crashes or errors found  - ...I think we're ready for the demo...
On a side note I discovered how to set the mouse speed, although it is systemwide not just for our application, so we'd have to remember the system setting if we want it to be different for just our game then reset it on form destroy. Something like that. Anyway, Windows API call is: SystemParametersInfo(SPI_SETMOUSESPEED,0,pointer(aValue),0);
Can also be SPI_GETMOUSESPEED in which case aValue will be filled with the mouse sensitivity. Apparently it ranges from 0..20. Sound familiar? ;)