Fixed the out of sync bug in multiplayer!!!! It all came down to something not being initialised which meant the player list was screwed up and different on both PCs. I think it was also causing some crashes in multiplayer and other weird behaviour. Now it works and stays beautifully in sync!!! I tested on 1 PC issuing lots of commands for both players (warriors, building, etc.) for 15 minutes and everything was still perfect!) WE NOW HAVE MULTIPLAYER BETTER THAN KAM TPR!!!!!!!!!!!!!!!! :D :D :D I also removed goals checks for now in multiplayer, we can add them later once we sort out alliances and stuff. It stops that annoying "YOU WIN" thing coming up each time you start a multiplayer game :P
I will run a test with 3 PCs tomorrow to see if it actually works. And can we please try playing over Hamachi some time? :)
Also, I'm discussing with my dad how to implement packet loss detection and error checking to the UDP packets. I'll talk to you about it some time. Sorry for not being online to chat tonight, I was busy with other stuff most of the time.