Major work on Linux compatibility - it's getting close!
Note: Linux can't handle: MyFunction().Value so I have to change that. For example, fViewPort.GetCenter.X does not work. This is very unfortunate as I think it is going to cause major issues later on.
Also, MAXWORD, MAXBYTE, etc. do not exist so I've changed them to high(word), high(byte) etc.
Please comment on the way I've done things if you're not happy with it.

For now I've disabled Windows dependant features like networking and full screen mode in Linux. We can find solutions to each one once it's compiling and the other features are working.