- Fights use the same damage system as KaM  - Mostly fixed the "first strike advantage" issues, needs further testing  - Watch towers now miss 20% of the time to be more similar to KaM (they were too deadly before)
This is still WIP, please don't judge it too harshly yet. It needs tweaking/testing, and there could be oversights in the system. But I believe it is currently closer to KaM than ever before, and with less issues like first strike advantage than before. I'll be working on it.