- Made speed-up keys into toggle buttons  - Restored multiplayer check for speed-up keys!  - Fixed 50x speed checkbox
Basically the fGameSpeed method takes hardcoded values (0-3) as parameters, and sets the speed accordingly. The reason the 50x speedup box stopped working was because I forgot that speed existed so I didn't hardcode it. Now the method will take any non-hardcoded values literally.

This may not be the best way for the method to work if we ever intend to add more speed values, but I didn't know how to access the speed constants in KM_Settings from KM_InterfaceGameplay, so I used hardcoded values instead. Hope this makes sense and is ok!