The beginnings of a "Dismiss unit" button. Everything works as intended so far, except that the confirmation screen only appears for a fraction of a second and I can't work out why (not urgent but help appreciated).
Also, should the button be available for Units, Warriors, or both? I've made it for both so far which has made parts of the code difficult to keep organised (i.e. it uses "Army_Issue_Command", since normally units cannot be issued commands).

Any thoughts about any of this, or how it should look or function? Really this is just a prototype so feel free to make any suggestions at all :)