Trying an idea that massively improves the effectiveness of BuildList job assignment. Currently it's only written for Fieldworks for testing. Adding others won't be hard but I wanted to get your opinion first.
Please read my explanation of the idea in comments in TKMBuildList.UpdateState.

To test, run Worker Test and build a field near each worker. You will see that the job will be taken by the worker you place it closest to. Now give the workers lots of jobs so they are all working. You will see as a worker finishes his job, he takes the next closest job to him.

Now try it again with my solution commented out so it uses only the old job assignment method. You will see that when you drop a road, ANY random worker could take the job. When all workers are busy it will function effectively, but not when most are idle.

This solution seems to work VERY well in my opinion. The code is a bit more complex but not too bad consider how much better it is. Please let me know what you think :)